Buttons: Bring up DU hardware key action binding framework
Single, double, and long press taps are supported. Actions options
are the same as Fling and SmartBar, which is basically anything.
Single tap back button and single tap home button are fix and
can not be changed.

We use the same config_deviceHardwareKeys overlay to get present
button bits. Default is Nexus device settings, which is volume
rocker only. Camera button action binding is not supported at
this time.

PS2: update for using System.Secure uri

Change-Id: Idb2a3380f09ac14c65ba1f7727b7000c7d6e3339