Wifi: Judge saved AP's security when update AP list
When the saved AP change its security, the AP list in wifi settings will
appear two APs with the same MAC but different security.

When update AP list, WifiTracker should judge the saved AP's security,
if the security is not match, the AP would not appear in the wifi
settings.

Change-Id: Id1c84b095c25d7d8bdd62c9080dc2e05b35d0e3f
CRs-Fixed: 2186427